Update logic corrected + file upload/transfer
Turns out that the logic for defining the SQL string was correct and that the execute script was incorrect. I was storing different placeholders and then attempting to execute whilst binding data to placeholders that didn't exist. File upload has been solved with help of FormData, next I need to store temp file, edit if statements to include file path, move file and store path in database