Showgroups now shows group members aswell as groups
If a group id has been set then an individual is viewing a single group. Information relevant to that group such as members must be accessed. For this to happen we must use the bridge table usergroups to make the connection between users and groups. We then grab information on the user and their assosiated group by joining tables in the SQL statement. The next script to do is Addgoals.php