Saving changes to Update script before testing
I understand this code is not DRY or optimised. It doesn't utilise functions or for loops. But at this stage I cannot afford to set aside the time for optimisation / re-thinking, I just need to ensure it works. As the performance difference should be marginal in this case (small app) and that it is a prototype, it should be fine!