Update goals script works now
The solution was that I was trying to access an assosiative array as a normal array, bypassing the index for each row and accessing column fields directly. To correct this I implemented the index for each result before the assositive reference. Also, the UPDATE syntax was corrected as previously it had thrown errors, PHP my admins SQL insert template helped tremendously