fixed validation - delete.php
I was checking if it was set and then applying the relevant queries, but my test form had 3 fields to test so they were all set. This meant that only the first isset statement and the code enclosed would run, if the set field was empty then it would return 400 despite not being the field you wanted to test. To fix it I simply added a not empty cause to each statement to ensure that the ID passed was the ID intended