showtasks.php coded - 1 issue
Shows tasks for user ID or task ID, but doesn't retrieve information on either user or parent group/goal. Need to create a join query to store profile name and parent goal / group name. Currently this is secondary priority to completing the rest of the scripts and building the front-end. I will re-visit it as the UI is built and make changes if necessary, much of the information may already be received through other API calls since it is a single page app and props can cascade information for re-use