Updated showtasks.php
Previouslly any user with basic knowledge of query strings could pass a task_id paramter for the showtasks script and view other users tasks. Atm I have it so that only group members can view the single task pages. Now that I think about it I need to add the user to this exception otherwise the user will be able to access a list of their tasks but not the individual pages.