showtasks works perfectly
Task owners can view their tasks, but non owners cannot. However, if that tasks parent is a group, then other group members can view that task. This basic logic dictates the difference between private / public in my application, separting tasks into group tasks and personal tasks. Both require a degree of authentication to view, whether that be matching userid to the currently logged in user or a matching record in the usergroups table.