Finished improvements to addtasks.php
I've included the parent goal into the add task functionality. We first use a SQL query to find the selected goal and it's assosiated ID. We then check the current user ID against the goals owner ID. We use this ID to make sure the user owns the goal as multiple users can have the same goal name. I've just identified another problem. I've linked the groupname the same as goalname. But multiple users can belong to a group and a group can have multiple tasks. So I need to link the query to the usergroups table somehow to differentiate owners and members.