Join group .php - added some logic
Discovered some additional logic that was necessary. For example, I need to check that the user doesn't already belong to that group, and that group members are less than the group size limit and that the group exists in the database table. I've setout the basic logic I want to implement, but there are some errors with a few of my evaluations, particularly comparing the group size against members. I will need to find a solution to this before proceeding, as group functionality is a must have feature.