leavegroup.php completed
The script removes the user from the usergroup table and subtracts members by 1. It works! Next script: Complete.php - the script that changes tasks, goals and groups from pending status to complete. This one script should read in two variables, yes and no. Based on these it can mark events as in-complete, or complete. Events that have yet to pass will remain as pending. I currently need to workout the logic behind listing the overdue events