Completed motivate/like script
If a user has liked a task then the script will remove the like, if a user has not liked a task they will add a like. To determine which user liked/motivated which task we use a bridge table similiar to how we connected users to groups. Each time a user motivates a users task they get added to the bridge table which connects their user id with the task id. The script uses this table to determine whether it should add or remove a motivate. When removing a motivate its important to remove the record of the user in the bridge table just as it is important to add them when adding a motivate