Started building the profile page
Building the page-profile vue component. Hvae already built the sidebar component but it doesn't include router links. After finishing the profile component I will replace the static anchor tags with router tags and setup routes to each page. At the moment I am liking the style and the scripts are working as they should. I encountered a problem with binding data to images and classes, but found that it was because I was attempting to bind mustache brackets and not just enter the name. You bind to a property which is referenced by template tags, you don't bind to the template tags