Altered delete script to delete foreign constraints
I encountered a problem when trying to delete a goal on the profile page. The goal wouldn't delete because not only did tasks belong to the goal, but task motivators belonged to the tasks that belong to the goal. So I needed to systematically delete each record from child constraint to parent whilst also allowing goals without children to be deleted as normal. a foreach loop allowed me to delete the children and the rest was left as normal.