Deleting works -- finally!
Initial solution included storing goal id in a variable using a for loop to compare the stored ID against the array of goals. My problem was that to store the goal ID I needed to pass it as data to the function or use an event object to target the ID of the clicked element. If I did this then the context of This would change to the passed arguement and not the vue data model, preventing me from comparing the passed goal id against the array. This meant that I could either pass the goal id to remove the goal from the database or splice the array, but not both. I tried to pass a external function as a parameter which returned the array and goal id, but this didn't work. I also tried including two functions in the click event but this also failed. In the end the simplest solution was the correct one. I used vanilla javascript and parentNode and ChildNode to remove the element which ID matched the stored ID. This problem has costed me alot of time and I should have priotised features earlier and skipped this.