Trying to add Vue-Events-Calendar
Installed vue events calendar, trying to customize events so that the events are retrieved from the tasks array and not demoevents. Demoevents uses column names like title and date, whereas my data includes start_date and name. I need to find a way to customize this template to fit my needs. But don't forget to comment the the code to give credit to the original plugin developer so that people do not misassosiate it as your own. The Vue-events calendar is to act as a fallback for event viewing functionality required by the specification. Originally I intended to develop my own calendar but I have ran out of time and this plugin is a better solution than not implementing the requirement