Goal page shows goals and groups
Changed get request to include mygroup parameter to show users groups, also edited the PHP script so thatusers can see which group they belong to and not just the owner of the group. I had an issue where I couldnt deleted groups. This was supported by the issue that all users saw all groups. By adding the mygroups query parameter I made it so that just the users groups where displayed. This meant that the user could delete them. previouslly the user was attempting to delete a group they didn't belong to or own which resulted in an unauthorized request. Now there is little chance of unauthorized requests by design