Corrected addgroup logic so that owners are added to usergroup table
Owners are members of a group, but previouslly they were not added as such. On creation the script registered that the owner was a member and set members to 1 but this was meaningless without actually adding the owner to the usergroup table as a recorded user. This problem became apparent after I edited the showgroups script to display members of groups aswell as owners, now an owner can see their own group, members can see their own group but only owners can delete the group. If a member of a group deletes the group they will be removed from the usergroup table, but the group will remain. The only user with authority to delete the main group record is the user with a matching ownerID