I hope I fixed showtasks for groups
Previouslly the SQL for showtasks displayed all tasks scheduled for today for users of that group. But for some reason it didn't count the group owner as a user of the group. It would recognise that they are a user in the members request, but when it came to the SQL statement in showtasks it wouldn't display the owners tasks. I have since changed it about a billion times trying different logic and one has clicked, I will continue to test against the success criteria and check that it works across multiple users