Tasks filtered by date + status
If todays date is greater than the due_date then tasks will ask you for feedback on status. Otherwise they will display the status of pending, as they have not yet taken place. THere is also ana dditional clause, that if the status is equal to complete that the status will be returned in place of the buttons. This makes the base logic as follows: for tasks that haven't taken place yet, show pending. If they hvae been marked as complete, show complete, otherwise if tasks due date has passed and not complete show links prompting user for feedback